very simple lsp wasm connection
